feat: update data types in verification functions and tests to use uint256

+++ src/Zkolar.sol
@@ -11,10 +11,10 @@ contract Zkolar {
     }
 
     function verifyCredential(
-        uint[2] memory a,
-        uint[2][2] memory b,
-        uint[2] memory c,
-        uint[1] memory publicSignal
+        uint256[2] memory a,
+        uint256[2][2] memory b,
+        uint256[2] memory c,
+        uint256[1] memory publicSignal
     ) public view returns (bool) {
         return verifier.verifyProof(a, b, c, publicSignal);
     }
